+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2
+
+* Added theme.
+* Added initial foundation for components (colors, text styles, theme).
+* Added avatar component.
+* Added Source Sans as main font.
+
 ## 0.0.1
 
 * Added README.md and README.pt-BR.md files.